Make sure that we don't pass bogus RsScriptCall data along to driver.
In the case where there is no actual RsScriptCall data, our rs.spec is
accidentally passing along a stale pointer (with 0 length). Unfortunately,
the RS HAL does not supply the RsScriptCall length along with the data,
making it impossible to determine whether or not the sc data is real or
bogus.

Note that this path only occurs when we are creating packed data in
our JNI layer due to having a usrData parameter too. In other cases,
we don't end up propagating the bogus pointer(s).

Change-Id: Ica66b75abb63bcf33d4d536076cf326b5e4d8338