Build NEON run-time when targetting cortex-a15
The build of the NEON run-time library has been disabled for cortex-a15, as
performance with NEON enabled was lower than performance without NEON. We now
have ideas why this was the case. To enable testing of NEON performance and
later to switch on NEON by default, we now enabled the build of the library.

Change-Id: Iaf44e221a124d7829425c47a7e36aa43dc2fd02d