Detect use of RS objects in kernels to disable a ScriptGroup optimization.
Bug: 10394865

This change disables a ScriptGroup optimization when we can't guarantee
that there is no implicit ordering between two kernels. In this case, it
is possible to communicate between kernels using the same bound global
RS object (like Allocation). A subsequent kernel in the ScriptGroup could
accidentally pick up stale data, leading to incorrect results.

Now, we disable this optimization whenever we see binds/sets of variables
that are potentiall RS object types for a given Script/kernel. This is
overly conservative, but sufficient for now.

This change also fixes a small issue with preLaunch/postLaunch missing for
the default ScriptGroup execution case.

Change-Id: I0d19d200cc8dc397d68008a4df6ea423b1e4d04f