Fix RsBalls kernel invocation.
Bug: 11413011

The balls.rs Script was passing the same allocation twice (once as input and
once as output) to rsForEach(). The underlying kernel actually only accepts
an output parameter (and no input parameter). This change replaces the
"ain" misnomer with a proper "aout", and uses "aNull" to signify an empty
(unused) input Allocation.

Change-Id: I0c7cbba51f47587e26bbf85bb9a3d79c0b7e4b19