Add tests for rsDebug() in support library and fix exported functions.
Functions using long vector types for rsDebug were being exported with the
wrong signature from libRSSupport.so. Essentially it comes down to the name
mangling rules being different for "long" and "long long" when combined with
the ext_vector_type attribute. We now hard-code the "long" vector types in
our type signatures and then have them be cast properly to the 64-bit types
we were expecting.

Change-Id: I8f3dad73d0f17f22f66703aff56d6f712fa6df1f