Fix incorrect dependence on the system-built liblog.
Our previous Android.mk was incorrectly pulling in the latest liblog,
instead of linking dynamically against the older prebuilt liblog. This
causes problems due to the latest liblog using __android_fatal(), which
is a new function in bionic.

This change also updates our tests to use the checked-in prebuilt liblog.so
instead of the system-built liblog.

Change-Id: I211ef3a31f033e66d46c619d89d507ea65ec3731