Improve CTS generated tests (rs/api).
Change the way we track the range of allowed values: track
minimum and maximum separately, handle NaN and infinity better,
allow error on input and output.

With these new changes, improve many individual tests.

Instead of setting a global specifying the test expectations,
instantiate an object that contains this info so that parallel
tests won't interfere.

Fix a bug for the rare test that disregards secondary output
when the primary output is NaN.

Change-Id: I9e822191713dd494c03ebe515710bd490463e68b