Return a void* from AllocationGetPointer!
Bug: 16846318

We mistakenly used a function that returned void with a function pointer
that expected to return back a void pointer. This leads to undefined
behavior, since it is unknown what value will be present in the space
where the return value would appear. Under the 64-bit C++ tests, this
was returning invalid pointers (due to optimizations moving things around).

Change-Id: Iea75a29212a06bbacc2fc523f9c43b3efe471a31