Add Get/SetElement variants to sync mangled names
We miss some symbols for rs{Get,Set}Element_long,ulong variants because
64 bit integer values are 'long' in RS-land but might be 'long long' in
the driver.   This causes mangled names in the driver to correspond to
'long long'.

Define native_long and native_ulong types to be vectors
of 'long' as seen by the driver and define overloaded versions of
rsSetElementAt_ and rsGetElementAt_.  This should get us the correct
mangled names in the driver.  Since the parameters are actually
pointers, there is no correctness issue.

Change-Id: Ib8f936a6ca39f6becf714abe4e8bc96c140bee0f