Fail if non-threadable calls are in a kernel
bug 19095896

Calling rsAllocationIoSend, rsAllocationIoReceive,
rsAllocationCopy1DRange or rsAllocationCopy2DRange in a kernel is bad.
For one, they are not threadsafe.  Calls to these functions in an
invokable are common, so we do not want to summarily mark scripts as not
threadable if these functions are called.  Instead, this patch to the
driver detects if these functions are called inside a kernel and if so,
sends a fatal error message.

Change-Id: I10d3ef06cb8ed19a6bde686e71092d14ea58f5ec