Add a way to link against different driver names.
The existing linker path hard-coded "-lRSDriver" into the command line, but
this won't work for partner drivers that have a different implementation
loaded. In order to still properly handle use of the CPU driver, this needs
to change depending on whether we actually loaded an OVERRIDE_RS_DRIVER or
not.

bug 20894664

Change-Id: I0c4a4f12f5db819b234952bc8f364ac6300f147b