Added in wait for attach context flag.
Added in a small flag which is  passed to the context from the Java/C++ API
which causes the context to delay execution until a debugger is a attached.
This gives time to the developer to attach a debugger to the process before
the kernel is executed.

Change-Id: I5bb8e75100436920f9e1a0f414f3b0bcf0d2b439
Signed-off-by: Stephen McGroarty <stephen@codeplay.com>
(cherry picked from commit d5164d537c84867880694327727b5f22b0a3ccbd)