Fix BLEND_DST_ATOP and disable vectorized optimizations temporarily.
Bug: 22047392

This function needed to assign the 'w' component (alpha channel) of the
source input to the output. This patch does that for the C++ code, but
needs to be updated for the inline assembly that we have for
NEON/ASIMD/SSE. For the time being, we can just disable those paths.

Change-Id: Iaac02ae2d2cb670d299446d5a017d518c0e303da
(cherry picked from commit 820e22b13fb48bc4719fa606ee4bc47e184970e0)