Ensure that SSE-specific functions get built separately for debug runtime.
Bug: 22530323

The build rules for the debug context don't allow the SSE optimized
versions of some functions to be used. Since we bundle all of these
functions into a single file, this results in the debug runtime missing
some symbols (clamp, length, dot, sqrt). This change ensures that the
debug runtime for x86 gets the generic definitions of these functions
instead of dropping them completely.

Change-Id: Idedfbb5c1badf0b88530a7e926dda6141443ea1f