Prepare for arm64 prebuilts of RS support library.
Bug: 22693954

This change fixes update_rs_prebuilts.sh to build for arm64 as well. It
also makes some changes to RSTest_CompatLib for working on arm64 (i.e.
removed APIs, etc.). The old version of RSTest_CompatLib is retained as
RSTest_CompatLibLegacy, but it is disabled from building due to errors
with mma and FORCE_BUILD_RS_COMPAT. We also now need to ensure we
properly build all the required libraries (both .so and .a files).

Change-Id: If89b8a542eb0d61db0ba186d4e3221bcc6e5664b