Remove RSYuvFormat from C++ API header
Remove definition of RSYuvFormat from the C++ API's headers and use the
newly added enum in rsDefines.h (which is distributed as a part of the
NDK).  The new enum is named 'RsYuvFormat' for sake of consistency with
the rest of rsDefines.h.

The new enum also uses the Camera's pixel formats as the enum values.
This requires minor updates.

Change-Id: Ia04e320c45d0e13cb476d2ad4deefeb2f8da60d3