Updated spec for the rsForEachInternal API
Bug: 23535985

Missed the spec for rsForEachInternal in previous CLs.
The right prototype is
rsForEachInternal(int, rs_script_call*, int, int, ...)
or mangled as
_Z17rsForEachInternaliP14rs_script_calliiz

The bcc whitelist was correctly updated in previous CLs.

This is an internal API, therefore header files are not affected.

Updated the api generator to handle the new case of ellipsis being used
in a C++/overloadable, name-mangled function.

Change-Id: Ib6e1341b02eccc64574ebca04c54a7673cb9726e