Remove the always_inline attribute on all RS builtins.
Remove the always_inline attribute on all RenderScript builtins
in the specification files. This leads to the removal of that
attribute from the declarations of all those functions in the
RenderScript header files.
Also, updates a comment in the unrelated file rs_for_each.rsh.
I presume someone forgot to update that when they changed the
corresponding spec file.

Signed-off-by: Verena Beckham <verena@codeplay.com>

Change-Id: Ia4cdffdbdc9866a5b961356218dd01732bf2e830