Avoid breaking on dead launch options
Bug: 28463187

Launch options array[2-4] are not initialized by the support library
in the current SDK.

They may have arbitrary values that invalidate the checks that this CL
is removing, and cause a kernel launch to bail out unnecessarily.

This CL removes those checks, since the array.*Start and array.*End
fields in the launch options are not used anywhere in the current
runtime.

Change-Id: Ib8a2e551619f770b402f569acbde635ee047675a